Changes 29/10/24

+++ CartItem.cs
@@ -9,7 +9,7 @@ namespace Townbush_Pharmacy_Website
     {
         public int ProdID { get; set; }
         public string ProdName {get; set; }
-        public int ProdPrice { get; set; }
+        public decimal ProdPrice { get; set; }
         public int Quantity { get; set; }
         public decimal SubTotal { get; set; }
         public decimal GetSubTotal()
